Started camera networking.
Mostly just setup for networking and tweaks to some existing scripts to support this.

There is one minor-ish bug that will be fixed in later builds as this is developed: When disconnection from mirror it loads the main scene again which destroys the camera. This will be fixed in the next commit or two. Workaround for now is to re-connnect to a server, then disconnect and the camera is recreated.